protocol: move cgo::{Out, Service} to protocol crate
Allow other crates to implement Services, without bringing in the
rest of the `bindings` crate.

This also allows `bindings` to depend on other crates (e.x. `derive`)
where service implementations live, and trivially bind them.
